tests/wpt/web-platform-tests/workers/constructors/SharedWorker/number-arguments.html

<!--
onconnect = function(e) {
  e.ports[0].postMessage(['FAIL', self.name]);
}
/*
-->
<!doctype html>
<title>1 as arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new SharedWorker(1, 1);
  worker.port.addEventListener('message', this.step_func(function(e) {
    assert_equals(e.data[0], '1', 'first arg');
    assert_equals(e.data[1], '1', 'second arg');
    this.done();
  }), false);
  worker.port.start();
});
</script>
<!--
*/
//-->
